pocmonad v0.1.0 Pocmonad.Happy

This is the Happy module, Ok, so I was just trying to find a nice way (beautiful syntax, yet flexible enough) to handle errors in elixir. Handling :ok/:error like tuples without lots of if/cases.

HEX

{:happy, "~> 1.3"}

Pros :

Clean to read

Its kind of with but less oppinied

Cons :

Implementation given for other people

Not follow the pipeline way

sounds like framework and not lib

Functions

try1()

input {:ok, any} call returns {:ok, any};

Examples

iex> Pocmonad.Happy.try1()
{:ok, 23}

try2()

input {:ok, any} call returns {:ok, any};

Examples

iex> Pocmonad.Happy.try1()
{:ok, 23}