pocmonad v0.1.0 Pocmonad.MonadEx

This is the MonadEx module, MonadEx introduces monads into Elixir. Monads encapsulate state and control the flow of code. A monad’s bind operation is similar, but more powerful, than Elixir’s built-in pipelines.

HEX

{:monadex, "~> 1.1"}

Pros :

There is more then only tagged tuple flow

maybe is other great way to avoid nullable issues

Popular solution

Cons :

Implementation given for other people

Strange simbol ~>>

big boilerplate

to much transformation

Functions

try1()

input any call returns {:ok, any};

Examples

iex> Pocmonad.MonadEx.try1()
{:ok, 23}

try2()

input any call returns {:error, Reason}.

Examples

iex> Pocmonad.MonadEx.try2()
{:error, :error_n}